+++ src/methods/get.js
@@ -1,11 +1,13 @@
 'use strict';
 
+const { isFunction } = require('../helpers/is');
+
 module.exports = function get(key, defaultValue = null) {
   if (this.items[key] !== undefined) {
     return this.items[key];
   }
 
-  if (typeof defaultValue === 'function') {
+  if (isFunction(defaultValue)) {
     return defaultValue();
   }