+++ src/methods/shift.js
@@ -1,13 +1,17 @@
 'use strict';
 
 module.exports = function shift() {
-  if (Array.isArray(this.items)) {
+  if (Array.isArray(this.items) && this.items.length) {
     return this.items.shift();
   }
 
-  const key = Object.keys(this.items)[0];
-  const value = this.items[key] || null;
-  delete this.items[key];
+  if (Object.keys(this.items).length) {
+    const key = Object.keys(this.items)[0];
+    const value = this.items[key];
+    delete this.items[key];
 
-  return value;
+    return value;
+  }
+
+  return null;
 };