+++ test/methods/whenNotEmpty_test.js
@@ -9,12 +9,38 @@ module.exports = (it, expect, collect) => {
     expect(collection.all()).to.eql([]);
   });
 
+  it('should return the default value when the collection is not empty', () => {
+    const collection = collect([1, 2, 3]);
+
+    collection.whenNotEmpty(c => c.push(4));
+
+    expect(collection.all()).to.eql([1, 2, 3, 4]);
+  });
+
+  it('should return the default function when the collection is not empty', () => {
+    const collection = collect([]);
+
+    collection.whenNotEmpty(c => c.push(4), c => c.push(5));
+
+    expect(collection.all()).to.eql([5]);
+  });
+
   it('should execute the callback when the collection object is not empty', () => {
     const collection = collect({});
 
     collection.whenNotEmpty(c => c.put('name', 'Mohamed Salah'));
 
     expect(collection.all()).to.eql({});
+  });
+
+  it('should execute the default function when the collection object is not empty', () => {
+    const collection = collect({});
+
+    collection.whenNotEmpty(c => c.put('name', 'Mohamed Salah'), c => c.put('name', 'Sadio Mané'));
+
+    expect(collection.all()).to.eql({
+      name: 'Sadio Mané',
+    });
   });
 
   it('should not execute the callback when the collection is not empty', () => {